requirements.txt
pygame==2.5.2
pytmx==3.32
pyaudio==0.2.14
cx_Freeze==6.15.16
web3==6.15.1
python-dotenv==1.0.0
openai>=1.0.0
anthropic>=0.8.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
